game_backend/requirements.txt
fastapi
uvicorn
supabase
gradio_client
python-multipart
python-dotenv
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
pydantic-settings
email-validator
python-dateutil
httpx
